@@ -19,25 +19,35 @@
 		"defaults"
 	],
 	"devDependencies": {
-		"@babel/core": "^7.12.10",
-		"@babel/preset-env": "^7.12.11",
-		"autoprefixer": "^10.2.3",
+		"@babel/core": "^7.13.8",
+		"@babel/eslint-parser": "^7.13.8",
+		"@babel/preset-env": "^7.13.9",
+		"@babel/preset-react": "^7.12.13",
+		"autoprefixer": "^10.2.4",
 		"babel-loader": "^8.2.2",
-		"browser-sync": "^2.26.13",
-		"core-js": "^3.8.3",
+		"browser-sync": "^2.26.14",
+		"core-js": "^3.9.1",
 		"cssnano": "^4.1.10",
+		"cross-env": "^7.0.3",
+		"eslint": "^7.21.0",
+		"eslint-config-prettier": "^8.1.0",
+		"eslint-plugin-import": "^2.22.1",
+		"eslint-plugin-jsx-a11y": "^6.4.1",
+		"eslint-plugin-react": "^7.22.0",
+		"eslint-plugin-react-hooks": "^4.2.0",
 		"npm-run-all": "^4.1.5",
-		"postcss": "^8.2.4",
+		"postcss": "^8.2.7",
 		"postcss-100vh-fix": "^1.0.2",
 		"postcss-cli": "^8.3.1",
 		"postcss-for": "^2.1.1",
 		"postcss-import": "^14.0.0",
 		"postcss-nested": "^4.2.3",
 		"postcss-scss": "^3.0.4",
-		"postcss-simple-vars": "^6.0.2",
-		"webpack": "^5.16.0",
+		"postcss-simple-vars": "^6.0.3",
+		"prettier": "2.2.1",
+		"webpack": "^5.24.3",
 		"webpack-build-notifier": "^2.1.1",
-		"webpack-cli": "^4.4.0"
+		"webpack-cli": "^4.5.0"
 	},
 	"dependencies": {
 		"@dogstudio/highway": "2.2.1",
@@ -49,9 +59,9 @@
 		"imagesloaded": "4.1.4",
 		"infinite-scroll": "3.0.6",
 		"ismobilejs": "1.1.1",
-		"locomotive-scroll": "4.0.6",
+		"locomotive-scroll": "4.1.0",
 		"photoswipe": "4.1.3",
-		"swiper": "6.4.7",
-		"tailwindcss": "^2.0.2"
+		"swiper": "6.4.15",
+		"tailwindcss": "^2.0.3"
 	}
 }
