@@ -1,12 +1,13 @@
 /* eslint-disable no-unused-vars */
-import { gsap } from 'gsap';
-import InfiniteScroll from './components/infiniteScroll';
-import Modal from './components/modal';
+import { gsap } from 'gsap'
+import InfiniteScroll from './components/infiniteScroll'
+import Modal from './components/modal'
 // import AjaxNavigation from './components/ajaxNavigation';
-import Slider from './components/slider';
-import Loader from './components/loader';
-import Toggle from './components/toggle';
+import Slider from './components/slider'
+import Loader from './components/loader'
+import Toggle from './components/toggle'
 // import SmoothScroll from './components/smoothScroll';
+<<<<<<< HEAD
 import Animations from './components/animations';
 import AnimationsFlip from './components/animationsFlip';
 import Cursor from './components/cursorV2';
@@ -14,21 +15,30 @@ import Search from './components/search';
 import Magnet from './components/magnetV2';
 import Billing from './components/billing';
 import GdprCompliance from './components/gdprCompliance';
+=======
+import Animations from './components/animations'
+import AnimationsFlip from './components/animationsFlip'
+import Cursor from './components/cursor'
+import Search from './components/search'
+import Magnet from './components/magnet'
+import Billing from './components/billing'
+import GdprCompliance from './components/gdprCompliance'
+>>>>>>> e29900efd0d965b82a9240cdf701ff06cc7f242b
 
-const infiniteScroll = new InfiniteScroll();
-const loader = new Loader();
-const modal = new Modal();
+const infiniteScroll = new InfiniteScroll()
+const loader = new Loader()
+const modal = new Modal()
 // const ajaxNavigation = new AjaxNavigation();
-const sliderPrimary = new Slider('.js-slider-primary');
-const toggle = new Toggle();
-const billing = new Billing();
+const sliderPrimary = new Slider('.js-slider-primary')
+const toggle = new Toggle()
+const billing = new Billing()
 // const smoothScroll = new SmoothScroll();
-const animations = new Animations();
-const animationsFlip = new AnimationsFlip();
-const cursor = new Cursor();
-const search = new Search();
-const magnet = new Magnet();
-const gdprCompliance = new GdprCompliance();
+const animations = new Animations()
+const animationsFlip = new AnimationsFlip()
+const cursor = new Cursor()
+const search = new Search()
+const magnet = new Magnet()
+const gdprCompliance = new GdprCompliance()
 
 const sliderGallery = new Slider('.js-slider-gallery', {
 	direction: 'horizontal',
@@ -47,7 +57,7 @@ const sliderGallery = new Slider('.js-slider-gallery', {
 			spaceBetween: 80,
 		},
 	},
-});
+})
 
 const carousel = new Slider('.js-carousel', {
 	loop: false,
@@ -65,34 +75,34 @@ const carousel = new Slider('.js-carousel', {
 	},
 	on: {
 		touchStart() {
-			const swiper = this;
+			const swiper = this
 			for (let i = 0; i < swiper.slides.length; i++) {
 				gsap.to(swiper.slides[i], {
 					duration: 0.6,
 					ease: 'circ.out',
 					scale: 0.9,
-				});
+				})
 				gsap.to(swiper.slides[i].querySelector('.carousel__media'), {
 					duration: 2,
 					ease: 'circ.out',
 					scale: 1.2,
-				});
+				})
 			}
 		},
 		touchEnd() {
-			const swiper = this;
+			const swiper = this
 			for (let i = 0; i < swiper.slides.length; i++) {
 				gsap.to(swiper.slides[i], {
 					duration: 0.2,
 					ease: 'circ.out',
 					scale: 1,
-				});
+				})
 				gsap.to(swiper.slides[i].querySelector('.carousel__media'), {
 					duration: 0.2,
 					ease: 'circ.out',
 					scale: 1,
-				});
+				})
 			}
 		},
 	},
-});
+})
