@@ -1,12 +1,12 @@
-import Utilities from '../utilities/utilities';
-import * as deepmerge from 'deepmerge';
-import { gsap } from 'gsap';
-import { ScrollTrigger } from 'gsap/ScrollTrigger';
-import { SplitText } from 'gsap/SplitText';
-import { DrawSVGPlugin } from 'gsap/DrawSVGPlugin';
-import { MorphSVGPlugin } from 'gsap/MorphSVGPlugin';
+import Utilities from '../utilities/utilities'
+import * as deepmerge from 'deepmerge'
+import { gsap } from 'gsap'
+import { ScrollTrigger } from 'gsap/ScrollTrigger'
+import { SplitText } from 'gsap/SplitText'
+import { DrawSVGPlugin } from 'gsap/DrawSVGPlugin'
+import { MorphSVGPlugin } from 'gsap/MorphSVGPlugin'
 
-gsap.registerPlugin(ScrollTrigger, SplitText, DrawSVGPlugin, MorphSVGPlugin);
+gsap.registerPlugin(ScrollTrigger, SplitText, DrawSVGPlugin, MorphSVGPlugin)
 
 export default class Animations {
 	/**
@@ -14,50 +14,50 @@ export default class Animations {
 	 * @param {Object} options - User options
 	 */
 	constructor(element, options) {
-		this.element = element || '[data-scroll]';
+		this.element = element || '[data-scroll]'
 		this.defaults = {
 			triggers: this.element,
-		};
+		}
 		this.DOM = {
 			html: document.documentElement,
 			body: document.body,
-		};
-		this.options = options ? deepmerge(this.defaults, options) : this.defaults;
-		this.updateEvents = this.updateEvents.bind(this);
+		}
+		this.options = options
+			? deepmerge(this.defaults, options)
+			: this.defaults
+		this.updateEvents = this.updateEvents.bind(this)
 
-		window.addEventListener('DOMContentLoaded', () => {});
+		window.addEventListener('DOMContentLoaded', () => {})
 
-		ScrollTrigger.addEventListener('scrollStart', () => {});
+		ScrollTrigger.addEventListener('scrollStart', () => {})
 
-		ScrollTrigger.addEventListener('scrollEnd', () => {});
+		ScrollTrigger.addEventListener('scrollEnd', () => {})
 
-		ScrollTrigger.addEventListener('refreshInit', () => {});
+		ScrollTrigger.addEventListener('refreshInit', () => {})
 
-		ScrollTrigger.addEventListener('refresh', () => {});
+		ScrollTrigger.addEventListener('refresh', () => {})
 
 		window.addEventListener('NAVIGATE_OUT', () => {
 			// ScrollTrigger.update();
 			// ScrollTrigger.refresh();
-		});
+		})
 
-		window.addEventListener('NAVIGATE_IN', () => {
-		});
+		window.addEventListener('NAVIGATE_IN', () => {})
 
-		window.addEventListener('NAVIGATE_END', () => {
-		});
+		window.addEventListener('NAVIGATE_END', () => {})
 
 		window.addEventListener('LOADER_COMPLETE', () => {
-			this.init();
-			this.initEvents(this.options.triggers);
-			Utilities.initObserver(this.options.triggers, this.updateEvents);
-		});
+			this.init()
+			this.initEvents(this.options.triggers)
+			Utilities.initObserver(this.options.triggers, this.updateEvents)
+		})
 	}
 
 	/**
 	 * Init
 	 */
 	init() {
-		this.DOM.element = document.querySelectorAll(this.element);
+		this.DOM.element = document.querySelectorAll(this.element)
 	}
 
 	/**
@@ -68,58 +68,43 @@ export default class Animations {
 		ScrollTrigger.defaults({
 			markers: false,
 			ease: 'power3',
-		});
+		})
 
 		gsap.utils.toArray(triggers).forEach((element) => {
 			if (element.dataset.scroll === 'js-splittext-chars') {
-				this.splitTextAnimationChars(element);
+				this.splitTextAnimationChars(element)
+			} else if (element.dataset.scroll === 'js-splittext-lines') {
+				this.splitTextAnimationLines(element)
+			} else if (element.dataset.scroll === 'js-rotation') {
+				this.rotationAnimation(element)
+			} else if (element.dataset.scroll === 'js-batch') {
+				this.batchAnimation(element)
+			} else if (element.dataset.scroll === 'js-scale') {
+				this.scaleAnimation(element)
+			} else if (element.dataset.scroll === 'js-draw-svg') {
+				this.drawSvgAnimation(element)
+			} else if (element.dataset.scroll === 'js-background-color') {
+				this.backgroundColorAnimation(element)
+			} else if (element.dataset.scroll === 'js-pin') {
+				this.pinAnimation(element)
+			} else if (element.dataset.scroll === 'js-pin-image-sequence') {
+				this.pinImageSequence(element)
+			} else if (element.dataset.scroll === 'js-pin-horizontal') {
+				this.pinHorizontalAnimation(element)
+			} else if (element.dataset.scroll === 'js-pin-horizontal-section') {
+				this.pinHorizontalSectionAnimation(element)
+			} else if (element.dataset.scroll === 'js-pin-vertical') {
+				this.pinVerticalAnimation(element)
+			} else if (element.dataset.scroll === 'js-comparison') {
+				this.comparisonAnimation(element)
+			} else if (element.dataset.scroll === 'js-parallax') {
+				this.parallaxAnimation(element)
+			} else if (element.dataset.scroll === 'js-video') {
+				this.videoAnimation(element)
+			} else {
+				this.defaultAnimation(element)
 			}
-			else if (element.dataset.scroll === 'js-splittext-lines') {
-				this.splitTextAnimationLines(element);
-			}
-			else if (element.dataset.scroll === 'js-rotation') {
-				this.rotationAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-batch') {
-				this.batchAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-scale') {
-				this.scaleAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-draw-svg') {
-				this.drawSvgAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-background-color') {
-				this.backgroundColorAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-pin') {
-				this.pinAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-pin-image-sequence') {
-				this.pinImageSequence(element);
-			}
-			else if (element.dataset.scroll === 'js-pin-horizontal') {
-				this.pinHorizontalAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-pin-horizontal-section') {
-				this.pinHorizontalSectionAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-pin-vertical') {
-				this.pinVerticalAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-comparison') {
-				this.comparisonAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-parallax') {
-				this.parallaxAnimation(element);
-			}
-			else if (element.dataset.scroll === 'js-video') {
-				this.videoAnimation(element);
-			}			
-			else {
-				this.defaultAnimation(element);
-			}
-		});
+		})
 	}
 
 	/**
@@ -127,67 +112,51 @@ export default class Animations {
 	 * @param {Object} target - New selector
 	 */
 	updateEvents(target) {
-		this.init();
+		this.init()
 		// console.log(target.dataset.scroll);
 
 		setTimeout(() => {
 			if (target.dataset.scroll === 'js-splittext-chars') {
-				this.splitTextAnimationChars(target);
+				this.splitTextAnimationChars(target)
+			} else if (target.dataset.scroll === 'js-splittext-lines') {
+				this.splitTextAnimationLines(target)
+			} else if (target.dataset.scroll === 'js-rotation') {
+				this.rotationAnimation(target)
+			} else if (target.dataset.scroll === 'js-batch') {
+				this.batchAnimation(target)
+			} else if (target.dataset.scroll === 'js-scale') {
+				this.scaleAnimation(target)
+			} else if (target.dataset.scroll === 'js-draw-svg') {
+				this.drawSvgAnimation(target)
+			} else if (target.dataset.scroll === 'js-background-color') {
+				this.backgroundColorAnimation(target)
+			} else if (target.dataset.scroll === 'js-pin') {
+				this.pinAnimation(target)
+			} else if (target.dataset.scroll === 'js-pin-image-sequence') {
+				this.pinImageSequence(target)
+			} else if (target.dataset.scroll === 'js-pin-horizontal') {
+				this.pinHorizontalAnimation(target)
+			} else if (target.dataset.scroll === 'js-pin-horizontal-section') {
+				this.pinHorizontalSectionAnimation(target)
+			} else if (target.dataset.scroll === 'js-pin-vertical') {
+				this.pinVerticalAnimation(target)
+			} else if (target.dataset.scroll === 'js-comparison') {
+				this.comparisonAnimation(target)
+			} else if (target.dataset.scroll === 'js-parallax') {
+				this.parallaxAnimation(target)
+			} else if (target.dataset.scroll === 'js-video') {
+				this.videoAnimation(element)
+			} else {
+				this.defaultAnimation(element)
 			}
-			else if (target.dataset.scroll === 'js-splittext-lines') {
-				this.splitTextAnimationLines(target);
-			}
-			else if (target.dataset.scroll === 'js-rotation') {
-				this.rotationAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-batch') {
-				this.batchAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-scale') {
-				this.scaleAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-draw-svg') {
-				this.drawSvgAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-background-color') {
-				this.backgroundColorAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-pin') {
-				this.pinAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-pin-image-sequence') {
-				this.pinImageSequence(target);
-			}
-			else if (target.dataset.scroll === 'js-pin-horizontal') {
-				this.pinHorizontalAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-pin-horizontal-section') {
-				this.pinHorizontalSectionAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-pin-vertical') {
-				this.pinVerticalAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-comparison') {
-				this.comparisonAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-parallax') {
-				this.parallaxAnimation(target);
-			}
-			else if (target.dataset.scroll === 'js-video') {
-				this.videoAnimation(element);
-			} 
-			else {
-				this.defaultAnimation(element);
-			}
-		}, 1000);
+		}, 1000)
 	}
 
 	/**
 	 * default Animation
-	*/
+	 */
 	defaultAnimation(item) {
-
-		const targetClass = item.dataset.scroll;
+		const targetClass = item.dataset.scroll
 
 		ScrollTrigger.create({
 			trigger: item,
@@ -196,22 +165,20 @@ export default class Animations {
 			toggleActions: 'play none none none',
 			// markers: true,
 			// once: true,
-		});
-
+		})
 	}
 
 	/**
 	 * splitText Animation Chars
-	*/
+	 */
 	splitTextAnimationChars(item) {
-		
-		gsap.set(item, { autoAlpha: 1 });
+		gsap.set(item, { autoAlpha: 1 })
 
 		const splitText = new SplitText(item, {
 			type: 'chars, words',
-		});
-		const target = splitText.chars;
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		})
+		const target = splitText.chars
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -221,7 +188,7 @@ export default class Animations {
 				scrub: targetScrub || false,
 				toggleActions: 'play none play reset',
 			},
-		});
+		})
 
 		tl.from(target, {
 			yPercent: 100,
@@ -229,21 +196,20 @@ export default class Animations {
 			stagger: 0.05,
 			duration: 0.5,
 			ease: 'back.inOut',
-		});
+		})
 	}
 
 	/**
 	 * splitText Animation Lines
-	*/
+	 */
 	splitTextAnimationLines(item) {
-
-		gsap.set(item, { autoAlpha: 1 });
+		gsap.set(item, { autoAlpha: 1 })
 
 		const splitText = new SplitText(item, {
 			type: 'lines',
-		});
-		const target = splitText.lines;
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		})
+		const target = splitText.lines
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -253,24 +219,23 @@ export default class Animations {
 				scrub: targetScrub || false,
 				toggleActions: 'play none play reset',
 			},
-		});
+		})
 
 		tl.from(target, {
 			y: 20,
 			opacity: 0,
 			stagger: 0.01,
 			duration: 0.01,
-		});
+		})
 	}
 
 	/**
 	 * rotation Animation
-	*/
+	 */
 	rotationAnimation(item) {
+		gsap.set(item, { autoAlpha: 1 })
 
-		gsap.set(item, { autoAlpha: 1 });
-
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 		const tl = gsap.timeline({
 			scrollTrigger: {
 				trigger: item,
@@ -279,45 +244,48 @@ export default class Animations {
 				scrub: targetScrub || false,
 				toggleActions: 'play none play reverse',
 			},
-		});
+		})
 
 		tl.from(item, {
 			x: 400,
 			rotation: 360,
-		});
+		})
 	}
-
 
 	/**
 	 * batch Animation
-	*/
+	 */
 	batchAnimation(item) {
+		gsap.set(item, { autoAlpha: 1 })
 
-		gsap.set(item, { autoAlpha: 1 });
-
-		const target = item.querySelectorAll('[data-scroll-target]');
+		const target = item.querySelectorAll('[data-scroll-target]')
 
 		// gsap.set(target, { y: 100, opacity: 0 });
 
 		ScrollTrigger.batch(target, {
 			interval: 0.1, // time window (in seconds) for batching to occur.
-			batchMax: 3,   // maximum batch size (targets)
-			onEnter: batch => gsap.to(batch, {autoAlpha: 1, stagger: 0.1, overwrite: true}),
-			onLeave: batch => gsap.set(batch, {autoAlpha: 0, overwrite: true}),
-			onEnterBack: batch => gsap.to(batch, {autoAlpha: 1, stagger: 0.1, overwrite: true}),
-			onLeaveBack: batch => gsap.set(batch, {autoAlpha: 0, overwrite: true})			
-		});
-		ScrollTrigger.addEventListener('refreshInit', () => gsap.set(target, { y: 0 }));
+			batchMax: 3, // maximum batch size (targets)
+			onEnter: (batch) =>
+				gsap.to(batch, { autoAlpha: 1, stagger: 0.1, overwrite: true }),
+			onLeave: (batch) =>
+				gsap.set(batch, { autoAlpha: 0, overwrite: true }),
+			onEnterBack: (batch) =>
+				gsap.to(batch, { autoAlpha: 1, stagger: 0.1, overwrite: true }),
+			onLeaveBack: (batch) =>
+				gsap.set(batch, { autoAlpha: 0, overwrite: true }),
+		})
+		ScrollTrigger.addEventListener('refreshInit', () =>
+			gsap.set(target, { y: 0 })
+		)
 	}
 
 	/**
 	 * scale Animation
-	*/
+	 */
 	scaleAnimation(item) {
+		gsap.set(item, { autoAlpha: 1 })
 
-		gsap.set(item, { autoAlpha: 1 });
-
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -327,19 +295,18 @@ export default class Animations {
 				scrub: targetScrub || false,
 				toggleActions: 'play none play reverse',
 			},
-		});
+		})
 
 		tl.to(item, {
 			scale: 1.5,
-		});
+		})
 	}
 
 	/**
 	 * parallax Animation
-	*/
+	 */
 	parallaxAnimation(item) {
-
-		gsap.set(item, { autoAlpha: 1 });
+		gsap.set(item, { autoAlpha: 1 })
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -347,21 +314,19 @@ export default class Animations {
 				toggleActions: 'play pause none none',
 				scrub: 2,
 			},
-		});
+		})
 
 		tl.to(item, {
 			y: -item.dataset.scrollSpeedY * 100 || 0,
 			x: -item.dataset.scrollSpeedX * 100 || 0,
-		});
+		})
 	}
-
 
 	/**
 	 * video Animation
-	*/
+	 */
 	videoAnimation(item) {
-
-		const targetVideo = item.querySelector('[data-scroll-target]');
+		const targetVideo = item.querySelector('[data-scroll-target]')
 
 		gsap.timeline({
 			scrollTrigger: {
@@ -371,28 +336,26 @@ export default class Animations {
 				markers: false,
 				onEnter: () => targetVideo.play(),
 				onLeave: () => {
-					targetVideo.pause();
-					targetVideo.currentTime = 0;
+					targetVideo.pause()
+					targetVideo.currentTime = 0
 				},
 				onEnterBack: () => targetVideo.play(),
 				onLeaveBack: () => {
-					targetVideo.pause();
-					targetVideo.currentTime = 0;
+					targetVideo.pause()
+					targetVideo.currentTime = 0
 				},
 			},
-		});
-
+		})
 	}
 
 	/**
 	 * darwSvg Animation
-	*/
+	 */
 	drawSvgAnimation(item) {
+		gsap.set(item, { autoAlpha: 1 })
 
-		gsap.set(item, { autoAlpha: 1 });
-
-		const target = item.querySelectorAll('path');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const target = item.querySelectorAll('path')
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -402,41 +365,41 @@ export default class Animations {
 				end: 'bottom 70%',
 				toggleActions: 'play none play reverse',
 			},
-		});
+		})
 
 		tl.from(target, {
 			drawSVG: 0,
-		});
+		})
 	}
 
 	/**
 	 * backgroundColor Animation
-	*/
+	 */
 	backgroundColorAnimation(item) {
-
-		const target = document.body;
+		const target = document.body
 		const tl = gsap.timeline({
 			scrollTrigger: {
 				trigger: item,
 				scrub: 1,
 				toggleActions: 'play reset play reset',
 			},
-		});
+		})
 
 		tl.to(target, {
 			backgroundColor: item.dataset.backgroundColor,
 			ease: 'power1',
-		});
+		})
 	}
 
 	/**
 	 * pin Animation
-	*/
+	 */
 	pinAnimation(item) {
-
-		const target = item.querySelector('[data-scroll-target]');
-		const targetElement = item.querySelectorAll('[data-scroll-target-animate]');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const target = item.querySelector('[data-scroll-target]')
+		const targetElement = item.querySelectorAll(
+			'[data-scroll-target-animate]'
+		)
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -449,45 +412,44 @@ export default class Animations {
 				scrub: targetScrub || false,
 				// anticipatePin: 1,
 			},
-		});
+		})
 
 		tl.from(targetElement, {
 			scale: 0.6,
 			transformOrigin: 'center center',
-		});
+		})
 	}
-
 
 	/**
 	 * pin Image Sequence https://codepen.io/GreenSock/pen/yLOVJxd
-	*/
+	 */
 	pinImageSequence(item) {
+		const target = item.querySelector('[data-scroll-target]')
+		const canvas = item.querySelector('[data-scroll-canvas]')
+		const targetContainer = item.querySelector(
+			'[data-scroll-target-animate]'
+		)
 
-		const target = item.querySelector("[data-scroll-target]");
-		const canvas = item.querySelector("[data-scroll-canvas]");
-		const targetContainer = item.querySelector("[data-scroll-target-animate]");	
+		const context = canvas.getContext('2d')
+		const frameCount = parseInt(item.dataset.scrollFrames, 10)
+		const framePath = item.dataset.scrollPath
 
-		const context = canvas.getContext("2d");
-		const frameCount = parseInt(item.dataset.scrollFrames, 10);
-		const framePath = item.dataset.scrollPath;
+		canvas.width = 900
+		canvas.height = 859
 
-		canvas.width = 900;
-		canvas.height = 859;
-
-		const currentFrame = index => (
+		const currentFrame = (index) =>
 			`${framePath}/${(index + 1).toString().padStart(4, '0')}.jpg`
-  			// `https://www.apple.com/105/media/us/airpods-pro/2019/1299e2f5_9206_4470_b28e_08307a42f19b/anim/sequence/large/01-hero-lightpass/${(index + 1).toString().padStart(4, '0')}.jpg`
-		);
+		// `https://www.apple.com/105/media/us/airpods-pro/2019/1299e2f5_9206_4470_b28e_08307a42f19b/anim/sequence/large/01-hero-lightpass/${(index + 1).toString().padStart(4, '0')}.jpg`
 
 		const images = []
 		const frames = {
-  			frame: 0
-		};
-		
+			frame: 0,
+		}
+
 		for (let i = 0; i < frameCount; i++) {
-			const img = new Image();
-			img.src = currentFrame(i);
-			images.push(img);
+			const img = new Image()
+			img.src = currentFrame(i)
+			images.push(img)
 		}
 
 		const tl = gsap.timeline({
@@ -499,30 +461,41 @@ export default class Animations {
 				pin: true,
 				// anticipatePin: 1,
 			},
-		});
+		})
 
-		tl	
-			.to(frames, { frame: frameCount - 1, snap: "frame", onUpdate: render, duration: 20},'together')
-			.fromTo(targetContainer, { scale:0.95 },{ scale:1, duration: 20},'together');
+		tl.to(
+			frames,
+			{
+				frame: frameCount - 1,
+				snap: 'frame',
+				onUpdate: render,
+				duration: 20,
+			},
+			'together'
+		).fromTo(
+			targetContainer,
+			{ scale: 0.95 },
+			{ scale: 1, duration: 20 },
+			'together'
+		)
 
-		images[0].onload = render;
+		images[0].onload = render
 
 		function render() {
-			context.clearRect(0, 0, canvas.width, canvas.height);
-			context.drawImage(images[frames.frame], 0, 0); 
+			context.clearRect(0, 0, canvas.width, canvas.height)
+			context.drawImage(images[frames.frame], 0, 0)
 		}
-		
 	}
-
 
 	/**
 	 * pin Horizontal Animation
-	*/
+	 */
 	pinHorizontalAnimation(item) {
-
-		const target = item.querySelector('[data-scroll-target]');
-		const targetContainer = item.querySelector('[data-scroll-target-animate]');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const target = item.querySelector('[data-scroll-target]')
+		const targetContainer = item.querySelector(
+			'[data-scroll-target-animate]'
+		)
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -538,88 +511,108 @@ export default class Animations {
 				// onEnterBack: () => {},
 				// onLeaveBack: () => {},
 			},
-		});
+		})
 
-		tl.fromTo(targetContainer, { x: 0 }, { x: -targetContainer.getBoundingClientRect().width + target.getBoundingClientRect().width });
+		tl.fromTo(
+			targetContainer,
+			{ x: 0 },
+			{
+				x:
+					-targetContainer.getBoundingClientRect().width +
+					target.getBoundingClientRect().width,
+			}
+		)
 	}
 
 	/**
 	 * pin Horizontal Section Animation
-	*/
+	 */
 	pinHorizontalSectionAnimation(item) {
+		const target = item.querySelector('[data-scroll-target]')
+		const section = item.querySelectorAll('[data-scroll-section]')
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
-		const target = item.querySelector('[data-scroll-target]');
-		const section = item.querySelectorAll('[data-scroll-section]');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const sections = gsap.utils.toArray(section)
 
-		const sections = gsap.utils.toArray(section);
-
-		let maxWidth = 0;
+		let maxWidth = 0
 
 		const getMaxWidth = () => {
-			maxWidth = 0;
+			maxWidth = 0
 			sections.forEach((section) => {
-				maxWidth += section.offsetWidth;
-			});
-		};
-		getMaxWidth();
-		ScrollTrigger.addEventListener("refreshInit", getMaxWidth);
+				maxWidth += section.offsetWidth
+			})
+		}
+		getMaxWidth()
+		ScrollTrigger.addEventListener('refreshInit', getMaxWidth)
 
 		gsap.to(sections, {
 			x: () => `-${maxWidth - window.innerWidth}`,
-			ease: "none",
+			ease: 'none',
 			scrollTrigger: {
 				trigger: target,
 				pin: true,
 				scrub: targetScrub || false,
 				start: 'center center',
 				end: () => `+=${maxWidth}`,
-				invalidateOnRefresh: true
-			}
-		});
+				invalidateOnRefresh: true,
+			},
+		})
 
 		// ADD SKEW
 		let proxy = { skew: 0 },
-		skewSetter = gsap.quickSetter(section, "skewX", "deg"), // fast
-		clamp = gsap.utils.clamp(-10, 10); // don't let the skew go beyond [X] degrees. 
+			skewSetter = gsap.quickSetter(section, 'skewX', 'deg'), // fast
+			clamp = gsap.utils.clamp(-10, 10) // don't let the skew go beyond [X] degrees.
 		// END SKEW
 
 		sections.forEach((sct, i) => {
 			ScrollTrigger.create({
 				trigger: sct,
-				start: () => 'top top-=' + (sct.offsetLeft - window.innerWidth/2) * (maxWidth / (maxWidth - window.innerWidth)),
-				end: () => '+=' + sct.offsetWidth * (maxWidth / (maxWidth - window.innerWidth)),
-				toggleClass: {targets: sct, className: "active"},
+				start: () =>
+					'top top-=' +
+					(sct.offsetLeft - window.innerWidth / 2) *
+						(maxWidth / (maxWidth - window.innerWidth)),
+				end: () =>
+					'+=' +
+					sct.offsetWidth *
+						(maxWidth / (maxWidth - window.innerWidth)),
+				toggleClass: { targets: sct, className: 'active' },
 				// ADD SKEW
 				onUpdate: (self) => {
-					let skew = clamp(self.getVelocity() / -500);
+					let skew = clamp(self.getVelocity() / -500)
 					// only do something if the skew is MORE severe. Remember, we're always tweening back to 0, so if the user slows their scrolling quickly, it's more natural to just let the tween handle that smoothly rather than jumping to the smaller skew.
 					if (Math.abs(skew) > Math.abs(proxy.skew)) {
-					  proxy.skew = skew;
-					  gsap.to(proxy, {skew: 0, duration: 0.5, ease: "circ", overwrite: true, onUpdate: () => skewSetter(proxy.skew)});
+						proxy.skew = skew
+						gsap.to(proxy, {
+							skew: 0,
+							duration: 0.5,
+							ease: 'circ',
+							overwrite: true,
+							onUpdate: () => skewSetter(proxy.skew),
+						})
 					}
-				}
+				},
 				// END SKEW
-			});
-		});
+			})
+		})
 
 		// SKEW: make the right edge "stick" to the scroll bar. force3D: true improves performance
-		gsap.set(section, {transformOrigin: "center center", force3D: true});
+		gsap.set(section, { transformOrigin: 'center center', force3D: true })
 		// END SKEW
-
 	}
-
 
 	/**
 	 * pin Vertical Animation
-	*/
+	 */
 	pinVerticalAnimation(item) {
-
-		const target = item.querySelector('[data-scroll-target]');
-		const targetLeft = item.querySelector('.js-pin-vertical-container-left');
-		const targetCenter = item.querySelector('.js-pin-vertical-container-center');
-		const targetRight = item.querySelector('.js-pin-vertical-container-right');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const target = item.querySelector('[data-scroll-target]')
+		const targetLeft = item.querySelector('.js-pin-vertical-container-left')
+		const targetCenter = item.querySelector(
+			'.js-pin-vertical-container-center'
+		)
+		const targetRight = item.querySelector(
+			'.js-pin-vertical-container-right'
+		)
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -630,21 +623,37 @@ export default class Animations {
 				pin: true,
 				// anticipatePin: 1,
 			},
-		});
-		tl.fromTo(targetLeft, { y: 0 }, { y: -targetLeft.offsetHeight + item.offsetHeight }, 'step')
-			.fromTo(targetCenter, { y: 0 }, { y: targetCenter.offsetHeight - item.offsetHeight }, 'step')
-			.fromTo(targetRight, { y: 0 }, { y: -targetRight.offsetHeight + item.offsetHeight }, 'step');
+		})
+		tl.fromTo(
+			targetLeft,
+			{ y: 0 },
+			{ y: -targetLeft.offsetHeight + item.offsetHeight },
+			'step'
+		)
+			.fromTo(
+				targetCenter,
+				{ y: 0 },
+				{ y: targetCenter.offsetHeight - item.offsetHeight },
+				'step'
+			)
+			.fromTo(
+				targetRight,
+				{ y: 0 },
+				{ y: -targetRight.offsetHeight + item.offsetHeight },
+				'step'
+			)
 	}
 
 	/**
 	 * comparison Animation
-	*/
+	 */
 	comparisonAnimation(item) {
-
-		const target = item.querySelector('[data-scroll-target]');
-		const targetMedia = item.querySelectorAll('.js-comparison-after-media');
-		const targetImage = item.querySelectorAll('.js-comparison-after-media img');
-		const targetScrub = parseInt(item.dataset.scrollScrub, 10);
+		const target = item.querySelector('[data-scroll-target]')
+		const targetMedia = item.querySelectorAll('.js-comparison-after-media')
+		const targetImage = item.querySelectorAll(
+			'.js-comparison-after-media img'
+		)
+		const targetScrub = parseInt(item.dataset.scrollScrub, 10)
 
 		const tl = gsap.timeline({
 			scrollTrigger: {
@@ -657,10 +666,10 @@ export default class Animations {
 				// anticipatePin: 1,
 			},
 			defaults: { ease: 'none' },
-		});
-			// animate the container one way...
+		})
+		// animate the container one way...
 		tl.fromTo(targetMedia, { xPercent: 100, x: 0 }, { xPercent: 0 })
 			// ...and the image the opposite way (at the same time)
-			.fromTo(targetImage, { xPercent: -100, x: 0 }, { xPercent: 0 }, 0);
+			.fromTo(targetImage, { xPercent: -100, x: 0 }, { xPercent: 0 }, 0)
 	}
 }
