@@ -2,10 +2,10 @@
  * Cursor module
  * Mouse interactions
  */
-import Utilities from '../utilities/utilities';
+import Utilities from '../utilities/utilities'
 
-const Deepmerge = require('deepmerge');
-const isMobile = require('ismobilejs');
+const Deepmerge = require('deepmerge')
+const isMobile = require('ismobilejs')
 
 export default class Cursor {
 	/**
@@ -13,40 +13,43 @@ export default class Cursor {
 	 * @param {Object} options - User options
 	 */
 	constructor(element, options) {
-		this.element = element || 'js-cursor';
+		this.element = element || 'js-cursor'
 		this.defaults = {
-			triggers: 'a, [type="submit"], .js-cursor-drag, .js-cursor-hover, .js-cursor-right, .js-cursor-left, .js-cursor-zoom, .js-cursor-close',
+			triggers:
+				'a, [type="submit"], .js-cursor-drag, .js-cursor-hover, .js-cursor-right, .js-cursor-left, .js-cursor-zoom, .js-cursor-close',
 			alwaysVisible: true,
 			cursorVisible: true,
 			interpolationAmount: {
 				inner: 0.15,
 				outer: 0.15,
 			},
-		};
-		this.options = options ? Deepmerge(this.defaults, options) : this.defaults;
-		this.updateEvents = this.updateEvents.bind(this);
-		this.getMousePosition = this.getMousePosition.bind(this);
-		this.click = this.click.bind(this);
-		this.toggle = this.toggle.bind(this);
-		this.render = this.render.bind(this);
-		this.scale = 1;
-		this.opacity = 1;
+		}
+		this.options = options
+			? Deepmerge(this.defaults, options)
+			: this.defaults
+		this.updateEvents = this.updateEvents.bind(this)
+		this.getMousePosition = this.getMousePosition.bind(this)
+		this.click = this.click.bind(this)
+		this.toggle = this.toggle.bind(this)
+		this.render = this.render.bind(this)
+		this.scale = 1
+		this.opacity = 1
 		this.mousePos = {
 			x: -100,
 			y: -100,
-		};
+		}
 		this.lastMousePos = {
 			x: 0,
 			y: 0,
-		};
-		this.lastScale = 1;
-		this.lastOpacity = 1;
+		}
+		this.lastScale = 1
+		this.lastOpacity = 1
 
 		window.addEventListener('DOMContentLoaded', () => {
-			this.init();
-			this.initEvents(this.options.triggers);
+			this.init()
+			this.initEvents(this.options.triggers)
 			// Utilities.initObserver(this.options.triggers, this.updateEvents);
-		});
+		})
 	}
 
 	/**
@@ -55,26 +58,26 @@ export default class Cursor {
 	init() {
 		this.DOM = {
 			element: document.getElementById(this.element),
-		};
+		}
 
 		if (!isMobile.any && this.DOM.element === null) {
-			return;
+			return
 		}
 
 		if (this.options.alwaysVisible) {
-			this.DOM.element.classList.add('is-cursor-always-visible');
+			this.DOM.element.classList.add('is-cursor-always-visible')
 		}
-		this.DOM.body = document.body;
+		this.DOM.body = document.body
 
 		if (!this.options.cursorVisible) {
-			this.DOM.body.style.cursor = 'none';
+			this.DOM.body.style.cursor = 'none'
 		}
-		this.DOM.outerCursor = document.getElementById('js-cursor-outer');
-		this.DOM.innerCursor = document.getElementById('js-cursor-inner');
-		this.DOM.icon = document.getElementById('js-cursor-icon');
-		this.bounds = this.DOM.element.getBoundingClientRect();
+		this.DOM.outerCursor = document.getElementById('js-cursor-outer')
+		this.DOM.innerCursor = document.getElementById('js-cursor-inner')
+		this.DOM.icon = document.getElementById('js-cursor-icon')
+		this.bounds = this.DOM.element.getBoundingClientRect()
 
-		this.renderID = window.requestAnimationFrame(this.render);
+		this.renderID = window.requestAnimationFrame(this.render)
 	}
 
 	/**
@@ -82,12 +85,12 @@ export default class Cursor {
 	 * @param {string} triggers - Selectors
 	 */
 	initEvents(triggers) {
-		window.addEventListener('mousemove', this.getMousePosition);
-		document.addEventListener('click', this.click);
-		[...document.querySelectorAll(triggers)].forEach((link) => {
-			link.addEventListener('mouseenter', this.toggle);
-			link.addEventListener('mouseleave', this.toggle);
-		});
+		window.addEventListener('mousemove', this.getMousePosition)
+		document.addEventListener('click', this.click)
+		;[...document.querySelectorAll(triggers)].forEach((link) => {
+			link.addEventListener('mouseenter', this.toggle)
+			link.addEventListener('mouseleave', this.toggle)
+		})
 	}
 
 	/**
@@ -95,9 +98,9 @@ export default class Cursor {
 	 * @param {Object} target - New selector
 	 */
 	updateEvents(target) {
-		this.init();
-		target.addEventListener('mouseenter', this.toggle);
-		target.addEventListener('mouseleave', this.toggle);
+		this.init()
+		target.addEventListener('mouseenter', this.toggle)
+		target.addEventListener('mouseleave', this.toggle)
 	}
 
 	/**
@@ -105,24 +108,24 @@ export default class Cursor {
 	 */
 	render() {
 		if (!isMobile.any && this.DOM.element === null) {
-			return;
+			return
 		}
 		this.lastMousePos.x = Utilities.lerp(
 			this.lastMousePos.x,
 			this.mousePos.x - this.bounds.width / 2,
-			this.options.interpolationAmount.outer,
-		);
+			this.options.interpolationAmount.outer
+		)
 		this.lastMousePos.y = Utilities.lerp(
 			this.lastMousePos.y,
 			this.mousePos.y - this.bounds.height / 2,
-			this.options.interpolationAmount.outer,
-		);
-		this.lastScale = Utilities.lerp(this.lastScale, this.scale, 0.15);
-		this.lastOpacity = Utilities.lerp(this.lastOpacity, this.opacity, 0.1);
-		this.DOM.element.style.transform = `translateX(${this.lastMousePos.x}px) translateY(${this.lastMousePos.y}px)`;
+			this.options.interpolationAmount.outer
+		)
+		this.lastScale = Utilities.lerp(this.lastScale, this.scale, 0.15)
+		this.lastOpacity = Utilities.lerp(this.lastOpacity, this.opacity, 0.1)
+		this.DOM.element.style.transform = `translateX(${this.lastMousePos.x}px) translateY(${this.lastMousePos.y}px)`
 		// this.DOM.outerCursor.style.transform = `scale(${this.lastScale})`;
-		this.DOM.outerCursor.style.opacity = this.lastOpacity;
-		this.renderID = window.requestAnimationFrame(() => this.render());
+		this.DOM.outerCursor.style.opacity = this.lastOpacity
+		this.renderID = window.requestAnimationFrame(() => this.render())
 	}
 
 	/**
@@ -131,18 +134,18 @@ export default class Cursor {
 	 */
 	click(event) {
 		if (!isMobile.any && this.DOM.element === null) {
-			return;
+			return
 		}
 		if (!event.target.classList.contains('js-cursor-drag')) {
-			this.DOM.element.classList.add('is-cursor-clicked');
+			this.DOM.element.classList.add('is-cursor-clicked')
 			setTimeout(() => {
-				this.DOM.element.classList.remove('is-cursor-clicked');
-			}, 100);
+				this.DOM.element.classList.remove('is-cursor-clicked')
+			}, 100)
 		}
 	}
 
 	getMousePosition(event) {
-		this.mousePos = Utilities.getMousePosition(event);
+		this.mousePos = Utilities.getMousePosition(event)
 	}
 
 	/**
@@ -151,56 +154,58 @@ export default class Cursor {
 	 */
 	toggle(event) {
 		if (!isMobile.any && this.DOM.element === null) {
-			return;
+			return
 		}
 
 		if (!event.target.classList.contains('js-cursor-drag')) {
-			this.DOM.element.classList.toggle('is-cursor-hover');
+			this.DOM.element.classList.toggle('is-cursor-hover')
 		}
 
 		if (event.target.classList.contains('js-cursor-drag')) {
-			this.DOM.element.classList.toggle('is-cursor-dragged');
+			this.DOM.element.classList.toggle('is-cursor-dragged')
 		}
 
 		if (event.target.classList.contains('js-cursor-zoom')) {
-			this.DOM.element.classList.toggle('is-cursor-zoom');
+			this.DOM.element.classList.toggle('is-cursor-zoom')
 		}
 
 		if (event.target.classList.contains('js-cursor-right')) {
-			this.DOM.element.classList.toggle('is-cursor-right');
+			this.DOM.element.classList.toggle('is-cursor-right')
 		}
 
 		if (event.target.classList.contains('js-cursor-left')) {
-			this.DOM.element.classList.toggle('is-cursor-left');
+			this.DOM.element.classList.toggle('is-cursor-left')
 		}
 
 		if (event.target.classList.contains('js-cursor-close')) {
-			this.DOM.element.classList.toggle('is-cursor-close');
+			this.DOM.element.classList.toggle('is-cursor-close')
 		}
 
 		if (event.target.classList.contains('js-cursor-plus')) {
-			this.DOM.element.classList.toggle('is-cursor-plus');
+			this.DOM.element.classList.toggle('is-cursor-plus')
 		}
 
 		if (event.target.classList.contains('js-cursor-hide')) {
-			this.DOM.element.classList.toggle('is-cursor-hide');
+			this.DOM.element.classList.toggle('is-cursor-hide')
 		}
 	}
 
 	destroy() {
 		if (!this.options.cursorVisible) {
-			this.DOM.body.style.cursor = '';
+			this.DOM.body.style.cursor = ''
 		}
 		this.mousePos = {
 			x: -100,
 			y: -100,
-		};
-		window.cancelAnimationFrame(this.renderID);
-		window.removeEventListener('mousemove', this.getMousePosition);
-		document.removeEventListener('click', this.click);
-		[...document.querySelectorAll(this.options.triggers)].forEach((link) => {
-			link.removeEventListener('mouseenter', this.toggle);
-			link.removeEventListener('mouseleave', this.toggle);
-		});
+		}
+		window.cancelAnimationFrame(this.renderID)
+		window.removeEventListener('mousemove', this.getMousePosition)
+		document.removeEventListener('click', this.click)
+		;[...document.querySelectorAll(this.options.triggers)].forEach(
+			(link) => {
+				link.removeEventListener('mouseenter', this.toggle)
+				link.removeEventListener('mouseleave', this.toggle)
+			}
+		)
 	}
 }
