@@ -1,6 +1,6 @@
-import Utilities from '../utilities/utilities';
-import * as deepmerge from 'deepmerge';
-import { gsap } from 'gsap';
+import Utilities from '../utilities/utilities'
+import * as deepmerge from 'deepmerge'
+import { gsap } from 'gsap'
 
 export default class Example {
 	/**
@@ -8,18 +8,20 @@ export default class Example {
 	 * @param {Object} options - User options
 	 */
 	constructor(element, options) {
-		this.element = element || '.js-example';
+		this.element = element || '.js-example'
 		this.defaults = {
 			triggers: this.element,
-		};
-		this.options = options ? deepmerge(this.defaults, options) : this.defaults;
-		this.updateEvents = this.updateEvents.bind(this);
+		}
+		this.options = options
+			? deepmerge(this.defaults, options)
+			: this.defaults
+		this.updateEvents = this.updateEvents.bind(this)
 
 		window.addEventListener('DOMContentLoaded', () => {
-			this.init();
-			this.initEvents(this.options.triggers);
-			Utilities.initObserver(this.options.triggers, this.updateEvents);
-		});
+			this.init()
+			this.initEvents(this.options.triggers)
+			Utilities.initObserver(this.options.triggers, this.updateEvents)
+		})
 	}
 
 	/**
@@ -28,8 +30,8 @@ export default class Example {
 	init() {
 		this.DOM = {
 			element: document.querySelectorAll(this.element),
-		};
-		this.DOM.child = document.querySelectorAll('.js-test-child');
+		}
+		this.DOM.child = document.querySelectorAll('.js-test-child')
 	}
 
 	/**
@@ -37,11 +39,12 @@ export default class Example {
 	 * @param {string} triggers - Selectors
 	 */
 	initEvents(triggers) {
-		this.example = 'Example';
-
-		[...document.querySelectorAll(triggers)].forEach((element) => {
-			element.addEventListener('mouseenter', () => { this.myAnimation(); });
-		});
+		this.example = 'Example'
+		;[...document.querySelectorAll(triggers)].forEach((element) => {
+			element.addEventListener('mouseenter', () => {
+				this.myAnimation()
+			})
+		})
 	}
 
 	/**
@@ -49,21 +52,25 @@ export default class Example {
 	 * @param {Object} target - New selector
 	 */
 	updateEvents(target) {
-		this.init();
-		target.addEventListener('mouseenter', () => { this.myAnimation(); });
+		this.init()
+		target.addEventListener('mouseenter', () => {
+			this.myAnimation()
+		})
 	}
 
 	/**
 	 * my Animation
-	*/
+	 */
 	myAnimation() {
-		console.log(this.example);
-
 		// Animation
-		const timelineExample = gsap.timeline({ delay: 0 });
+		const timelineExample = gsap.timeline({ delay: 0 })
 
 		timelineExample
-			.fromTo(this.DOM.element, { opacity: 1 }, { duration: 0.5, opacity: 0.2 })
-			.fromTo(this.DOM.child, { scale: 1 }, { duration: 0.5, scale: 3 });
+			.fromTo(
+				this.DOM.element,
+				{ opacity: 1 },
+				{ duration: 0.5, opacity: 0.2 }
+			)
+			.fromTo(this.DOM.child, { scale: 1 }, { duration: 0.5, scale: 3 })
 	}
 }
