@@ -3,10 +3,10 @@
  * Mouse interactions
  */
 
-import Utilities from '../utilities/utilities';
-import { gsap } from 'gsap';
+import Utilities from '../utilities/utilities'
+import { gsap } from 'gsap'
 
-const Deepmerge = require('deepmerge');
+const Deepmerge = require('deepmerge')
 
 export default class Magnet {
 	/**
@@ -14,27 +14,28 @@ export default class Magnet {
 	 * @param {Object} options - User options
 	 */
 	constructor(element, options) {
-		this.element = element || '.js-magnet';
+		this.element = element || '.js-magnet'
 		this.defaults = {
 			triggers: this.element,
-		};
+		}
 		this.DOM = {
 			html: document.documentElement,
 			body: document.body,
-		};
-		this.options = options ? Deepmerge(this.defaults, options) : this.defaults;
-		this.updateEvents = this.updateEvents.bind(this);
+		}
+		this.options = options
+			? Deepmerge(this.defaults, options)
+			: this.defaults
+		this.updateEvents = this.updateEvents.bind(this)
 
 		window.addEventListener('DOMContentLoaded', () => {
-			this.init();
-			this.initEvents(this.options.triggers);
-			Utilities.initObserver(this.options.triggers, this.updateEvents);
-		});
+			this.init()
+			this.initEvents(this.options.triggers)
+			Utilities.initObserver(this.options.triggers, this.updateEvents)
+		})
 	}
 
-
 	init() {
-		this.DOM.element = document.querySelectorAll(this.element);
+		this.DOM.element = document.querySelectorAll(this.element)
 	}
 
 	/**
@@ -42,11 +43,11 @@ export default class Magnet {
 	 * @param {string} triggers - Selectors
 	 */
 	initEvents(triggers) {
-		[...document.querySelectorAll(triggers)].forEach((magnet) => {
-			magnet.addEventListener('mousemove', this.moveMagnet);
+		;[...document.querySelectorAll(triggers)].forEach((magnet) => {
+			magnet.addEventListener('mousemove', this.moveMagnet)
 			magnet.addEventListener('mouseout', (event) => {
 				if (magnet.classList.contains('is-magnet')) {
-					magnet.classList.remove('is-magnet');
+					magnet.classList.remove('is-magnet')
 				}
 
 				gsap.to(event.currentTarget, {
@@ -54,21 +55,20 @@ export default class Magnet {
 					x: 0,
 					y: 0,
 					ease: 'back.out',
-				});
-			});
-		});
+				})
+			})
+		})
 	}
-
 
 	/**
 	 * Update events
 	 * @param {string} target - New selector
 	 */
 	updateEvents(target) {
-		target.addEventListener('mousemove', this.moveMagnet);
+		target.addEventListener('mousemove', this.moveMagnet)
 		target.addEventListener('mouseout', (event) => {
 			if (target.classList.contains('is-magnet')) {
-				target.classList.remove('is-magnet');
+				target.classList.remove('is-magnet')
 			}
 
 			gsap.to(event.currentTarget, {
@@ -76,29 +76,34 @@ export default class Magnet {
 				x: 0,
 				y: 0,
 				ease: 'back.out',
-			});
-		});
+			})
+		})
 	}
-
 
 	/**
 	 * moveMagnet
-	*/
+	 */
 	moveMagnet(event) {
-		const magnetButton = event.currentTarget;
-		const bounding = magnetButton.getBoundingClientRect();
+		const magnetButton = event.currentTarget
+		const bounding = magnetButton.getBoundingClientRect()
 
 		if (!magnetButton.classList.contains('is-magnet')) {
-			magnetButton.classList.add('is-magnet');
+			magnetButton.classList.add('is-magnet')
 		}
 		// console.log(magnetButton, bounding)
 
 		gsap.to(magnetButton, {
 			duration: 1,
-			x: (((event.clientX - bounding.left) / magnetButton.offsetWidth) - 0.5) * 70,
-			y: (((event.clientY - bounding.top) / magnetButton.offsetHeight) - 0.5) * 70,
+			x:
+				((event.clientX - bounding.left) / magnetButton.offsetWidth -
+					0.5) *
+				70,
+			y:
+				((event.clientY - bounding.top) / magnetButton.offsetHeight -
+					0.5) *
+				70,
 			ease: 'power4.out',
-		});
+		})
 
 		// magnetButton.style.transform = 'translate(' + (((( event.clientX - bounding.left)/(magnetButton.offsetWidth))) - 0.5) * strength + 'px,'+ (((( event.clientY - bounding.top)/(magnetButton.offsetHeight))) - 0.5) * strength + 'px)';
 	}
