@@ -4,11 +4,11 @@
  * @see https://github.com/locomotivemtl/locomotive-scroll
  */
 
-import Utilities from '../utilities/utilities';
-import LocomotiveScroll from 'locomotive-scroll';
-import * as deepmerge from 'deepmerge';
+import Utilities from '../utilities/utilities'
+import LocomotiveScroll from 'locomotive-scroll'
+import * as deepmerge from 'deepmerge'
 
-const imagesLoaded = require('imagesloaded');
+const imagesLoaded = require('imagesloaded')
 
 export default class SmoothScroll {
 	/**
@@ -20,7 +20,7 @@ export default class SmoothScroll {
 			body: document.body,
 			element: document.getElementById('js-scroll'),
 			scrollProgress: document.getElementById('js-scroll-progress'),
-		};
+		}
 		this.defaults = {
 			el: this.DOM.element,
 			elMobile: document,
@@ -43,49 +43,53 @@ export default class SmoothScroll {
 			touchMultiplier: 2,
 			triggers: '[data-scroll-to]',
 			updaterClass: 'js-scroll-update',
-		};
-		this.options = options ? deepmerge(this.defaults, options) : this.defaults;
+		}
+		this.options = options
+			? deepmerge(this.defaults, options)
+			: this.defaults
 
-		this.updateEvents = this.updateEvents.bind(this);
+		this.updateEvents = this.updateEvents.bind(this)
 
 		imagesLoaded(this.DOM.body, { background: true }, () => {
-			this.init();
-			this.initEvents(this.options.updater);
+			this.init()
+			this.initEvents(this.options.updater)
 			// TODO: Only 1 observer
-			Utilities.initObserver(this.options.triggers, this.updateEvents);
-			Utilities.initObserver('.' + this.options.updaterClass, this.updateEvents);
-		});
+			Utilities.initObserver(this.options.triggers, this.updateEvents)
+			Utilities.initObserver(
+				'.' + this.options.updaterClass,
+				this.updateEvents
+			)
+		})
 
 		window.addEventListener('NAVIGATE_OUT', () => {
-			this.stop();
-		});
+			this.stop()
+		})
 
 		window.addEventListener('NAVIGATE_IN', () => {
-			this.destroy();
-		});
+			this.destroy()
+		})
 
 		window.addEventListener('NAVIGATE_END', () => {
 			if (this.options.smooth) {
-				this.reinitialize();
-				this.progressBar();
-				this.direction();
+				this.reinitialize()
+				this.progressBar()
+				this.direction()
 			} else {
-				this.reinitialize();
+				this.reinitialize()
 			}
-		});
+		})
 	}
 
 	init() {
 		if (this.DOM.element.length === 0) {
-			return;
+			return
 		}
 
 		setTimeout(() => {
-			this.scroll = new LocomotiveScroll(this.options);
-			this.progressBar();
-			this.direction();
-		}, 200);
-
+			this.scroll = new LocomotiveScroll(this.options)
+			this.progressBar()
+			this.direction()
+		}, 200)
 	}
 
 	/**
@@ -93,18 +97,20 @@ export default class SmoothScroll {
 	 * @param {string} triggers - Selectors
 	 */
 	initEvents(triggers) {
-		const elements = document.querySelectorAll(triggers);
+		const elements = document.querySelectorAll(triggers)
 		for (let i = 0; i < elements.length; i++) {
 			elements[i].addEventListener('click', () => {
 				this.update()
-			});
+			})
 		}
 
-		const updaters = document.querySelectorAll('.' + this.options.updaterClass);
+		const updaters = document.querySelectorAll(
+			'.' + this.options.updaterClass
+		)
 		for (let i = 0; i < updaters.length; i++) {
 			updaters[i].addEventListener('click', () => {
 				this.update()
-			});
+			})
 		}
 	}
 
@@ -113,15 +119,15 @@ export default class SmoothScroll {
 	 * @param {string} target - New selector
 	 */
 	updateEvents(target) {
-		if ( target.classList.contains(this.options.updaterClass) ) {
+		if (target.classList.contains(this.options.updaterClass)) {
 			target.addEventListener('click', () => {
-				this.update();
-			});
+				this.update()
+			})
 		} else {
 			target.addEventListener('click', (event) => {
-				const href = target.getAttribute('href');
-				this.scrollTo(href);
-			});
+				const href = target.getAttribute('href')
+				this.scrollTo(href)
+			})
 		}
 	}
 
@@ -133,17 +139,20 @@ export default class SmoothScroll {
 		 * @param {Object} istance (delta, direction, limit, scroll, speed)
 		 */
 
-		if ( this.DOM.scrollProgress === null ) {
-			return;
+		if (this.DOM.scrollProgress === null) {
+			return
 		}
 
-		this.DOM.scrollProgress.style.height = 0;
+		this.DOM.scrollProgress.style.height = 0
 
 		this.on('scroll', (instance) => {
-			const progress = (instance.scroll.y / (this.DOM.element.offsetHeight - window.innerHeight)) * 100;
-			const progressRounded = Math.round(progress);
-			this.DOM.scrollProgress.style.height = `${progressRounded}%`;
-		});
+			const progress =
+				(instance.scroll.y /
+					(this.DOM.element.offsetHeight - window.innerHeight)) *
+				100
+			const progressRounded = Math.round(progress)
+			this.DOM.scrollProgress.style.height = `${progressRounded}%`
+		})
 	}
 
 	/**
@@ -154,39 +163,42 @@ export default class SmoothScroll {
 		 * @param {Object} istance (delta, direction, limit, scroll, speed)
 		 */
 
-		document.documentElement.setAttribute('data-direction', 'up');
+		document.documentElement.setAttribute('data-direction', 'up')
 
 		this.on('scroll', (instance) => {
-			document.documentElement.setAttribute('data-direction', instance.direction);
-		});
+			document.documentElement.setAttribute(
+				'data-direction',
+				instance.direction
+			)
+		})
 	}
 
 	/**
 	 * Restarts the scroll events
 	 */
 	start() {
-		this.scroll.start();
+		this.scroll.start()
 	}
 
 	/**
 	 * Stops the scroll events
 	 */
 	stop() {
-		this.scroll.stop();
+		this.scroll.stop()
 	}
 
 	/**
 	 * Reinitializes the scroll
 	 */
 	reinitialize() {
-		this.scroll.init();
+		this.scroll.init()
 	}
 
 	/**
 	 * Manually refresh the DOM
 	 */
 	update() {
-		this.scroll.update();
+		this.scroll.update()
 	}
 
 	/**
@@ -195,20 +207,20 @@ export default class SmoothScroll {
 	 * @param {number} offset
 	 */
 	scrollTo(target, offset) {
-		this.scroll.scrollTo(target, offset);
+		this.scroll.scrollTo(target, offset)
 	}
 
 	/**
 	 * On instance
 	 */
 	on(eventName, func) {
-		this.scroll.on(eventName, func);
+		this.scroll.on(eventName, func)
 	}
 
 	/**
 	 * Destroy instance
 	 */
 	destroy() {
-		this.scroll.destroy();
+		this.scroll.destroy()
 	}
 }
