@@ -1,7 +1,7 @@
 /**
  * Event dispatcher
  */
-import { DEBUG_MODE } from './environment';
+import { DEBUG_MODE } from './environment'
 
 export default class Dispatcher {
 	/**
@@ -10,11 +10,11 @@ export default class Dispatcher {
 	 * @param {*} detail - Additional parameters to pass along to the event handler.
 	 */
 	static trigger(eventName, detail) {
-		const event = new window.CustomEvent(eventName, { detail });
-		window.dispatchEvent(event);
+		const event = new window.CustomEvent(eventName, { detail })
+		window.dispatchEvent(event)
 		if (DEBUG_MODE) {
 			// eslint-disable-next-line no-console
-			console.log('🚀 Triggered:', eventName);
+			console.log('🚀 Triggered:', eventName)
 		}
 	}
 }
