@@ -10,7 +10,9 @@ export default class Validation {
 	 */
 	static isEmail(value) {
 		// eslint-disable-next-line max-len
-		return /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/.test(value);
+		return /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/.test(
+			value
+		)
 	}
 
 	/**
@@ -20,7 +22,9 @@ export default class Validation {
 	 */
 	static isUrl(value) {
 		// eslint-disable-next-line max-len
-		return /^(?:(?:(?:https?|ftp):)?\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})).?)(?::\d{2,5})?(?:[/?#]\S*)?$/i.test(value);
+		return /^(?:(?:(?:https?|ftp):)?\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})).?)(?::\d{2,5})?(?:[/?#]\S*)?$/i.test(
+			value
+		)
 	}
 
 	/**
@@ -29,7 +33,7 @@ export default class Validation {
 	 * @returns {boolean}
 	 */
 	static isInteger(value) {
-		return /^-?\d+$/.test(value);
+		return /^-?\d+$/.test(value)
 	}
 
 	/**
@@ -38,7 +42,7 @@ export default class Validation {
 	 * @returns {boolean}
 	 */
 	static isNumeric(value) {
-		return !isNaN(parseFloat(value)) && isFinite(value);
+		return !isNaN(parseFloat(value)) && isFinite(value)
 	}
 
 	/**
@@ -47,7 +51,7 @@ export default class Validation {
 	 * @returns {boolean}
 	 */
 	static isFloat(value) {
-		return /-?\d+\.\d+/.test(value);
+		return /-?\d+\.\d+/.test(value)
 	}
 
 	/**
@@ -56,6 +60,6 @@ export default class Validation {
 	 * @returns {boolean}
 	 */
 	static isEmpty(value, ignoreWhiteSpace) {
-		return (ignoreWhiteSpace ? value.trim().length : value.length) === 0;
+		return (ignoreWhiteSpace ? value.trim().length : value.length) === 0
 	}
 }
